third_party/WebKit/LayoutTests/platform/win/external/wpt/IndexedDB/idbobjectstore_add12-expected.txt

This is a testharness.js-based test.
PASS IDBObjectStore.add() - Attempt to add a record where the record's in-line key is not defined  
Harness: the test ran to completion.
